refactor: split monolithic file into modular components
- Separate main function into main.cpp
- Extract LineDetectorNode class to its own files
- Extract CameraCalibrator class to its own files
- Move all headers to src/ directory for better organization
- Fix vector parameter initialization for ROS 2 compatibility
- Update CMakeLists.txt to build with multiple source files

This modular structure improves code maintainability and makes
each component's responsibilities clearer.
